sockets: socket listener

+++ src/lib/runtime.rs
@@ -3,7 +3,7 @@ use std::path::PathBuf;
 pub const DEFAULT_RUNTIME_PATH: &'static str = "/tmp/stopwatchd";
 pub const DEFAULT_PIDFILE_PATH: &'static str = "/tmp/stopwatchd/pidfile";
 
-pub fn server_socker_path(pid: Option<usize>) -> PathBuf {
+pub fn server_socker_path(pid: Option<u32>) -> PathBuf {
     let path = PathBuf::from(DEFAULT_RUNTIME_PATH);
     match pid {
         Some(pid) => path.join(format!("server_socket_{}", pid)),
